chore, feat(#4): 대시보드 통계 APi 구현
- 패키지 구조 변경
  (OpenFeign 관련 클래스 external 패키지로 통일)
- 대시보드 통계 API 구현
