+++ src/main/java/com/modive/adminservice/api/dashboard/controller/AdminDashboardController.java
@@ -1,5 +1,6 @@
 package com.modive.adminservice.api.dashboard.controller;
 
+import com.modive.adminservice.api.dashboard.dto.res.TotalCntAndRateItem;
 import com.modive.adminservice.api.dashboard.dto.res.TotalEventCntByReasonItem;
 import com.modive.adminservice.domain.event.service.EventService;
 import com.modive.adminservice.global.dto.res.CommonRes;
@@ -40,10 +41,30 @@ public class AdminDashboardController {
         List<TotalEventCntByReasonItem> items = eventService.getTotalEventCntByType();
 
         Map<String, Object> data = new HashMap<>();
-        data.put("monthlyDrivesStatistics", items);
+        data.put("eventsStatisticsByReason", items);
 
         return new ResponseEntity<>(
-                CommonRes.success(data, "monthlyDrivesStatistics"),
+                CommonRes.success(data, "이벤트 유형별 추이 조회에 성공하였습니다."),
+                HttpStatus.OK
+        );
+    }
+
+    @GetMapping("/summary")
+    @Operation(summary = "대시 보드 통계", description = "대시 보드 상단의 통계 데이터를 조회합니다.")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Success",
+                    content = {@Content(schema = @Schema(implementation = CommonRes.class))}),
+            @ApiResponse(responseCode = "500", description = "Internal Server Error",
+                    content = {@Content(schema = @Schema(implementation = ErrorRes.class))})
+    })
+    public ResponseEntity<CommonRes> getMonthly () {
+        Map<String, TotalCntAndRateItem> dashboardStatistics = new HashMap<>();
+
+        Map<String, Object> data = new HashMap<>();
+        data.put("dashboardStatistics", dashboardStatistics);
+
+        return new ResponseEntity<>(
+                CommonRes.success(data, "대시 보드 통계 조회에 성공하였습니다."),
                 HttpStatus.OK
         );
     }
