+++ src/main/java/com/modive/adminservice/domain/event/repository/EventRepository.java
@@ -12,13 +12,13 @@ import java.util.List;
 
 @Repository
 public interface EventRepository extends JpaRepository<Event, Long> {
-    @Query("SELECT new com.modive.adminservice.event.dto.EventsByDriveDTO(et.eventTypeName, COUNT(e)) " +
+    @Query("SELECT new com.modive.adminservice.domain.event.dto.EventsByDriveDTO(et.eventTypeName, COUNT(e)) " +
             "FROM Event e INNER JOIN e.eventType et " +
             "WHERE e.driveId = :driveId " +
             "GROUP BY et.eventTypeId, et.eventTypeName")
     List<EventsByDriveDTO> countByTypeGroupedByDriveId(@Param("driveId") Long driveId);
 
-    @Query("SELECT new com.modive.adminservice.event.dto.EventTotalCntByTypeDTO(et.eventTypeName, COUNT(e)) " +
+    @Query("SELECT new com.modive.adminservice.domain.event.dto.EventTotalCntByTypeDTO(et.eventTypeName, COUNT(e)) " +
             "FROM Event e JOIN e.eventType et " +
             "GROUP BY et.eventTypeId")
     List<EventTotalCntByTypeDTO> totalCntByType();
