--- src/main/java/com/modive/adminservice/external/client/user/dto/res/UCUserListResData.java
@@ -1,23 +1,0 @@
-package com.modive.adminservice.external.client.user.dto.res;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.List;
-
-/**
- * 유저 서비스의
- * "/user/list", "/user/filter" API 응답의 data 필드를 감싸는 DTO.
- * - 실제 응답 구조: { "status": ..., "message": ..., "data": { "users": [...] } }
- * - 사용처: UserClient
- */
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UCUserListResData {
-    private List<UCUserListItem> users;
-}
