+++ src/main/java/com/modive/adminservice/external/dashboard/dto/res/DCTotalCntAndRateItem.java
@@ -1,0 +1,14 @@
+package com.modive.adminservice.external.dashboard.dto.res;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class DCTotalCntAndRateItem {
+    private Long value;
+    private double changeRate;}
