+++ src/main/java/com/modive/adminservice/external/dashboard/dto/res/DCTotalDriveResData.java
@@ -1,0 +1,14 @@
+package com.modive.adminservice.external.dashboard.dto.res;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class DCTotalDriveResData {
+    private DCTotalCntAndRateItem totalDrives;
+}
