+++ src/main/java/com/modive/adminservice/external/user/dto/res/UCTotalCntAndRateItem.java
@@ -1,0 +1,13 @@
+package com.modive.adminservice.external.user.dto.res;
+
+import lombok.*;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UCTotalCntAndRateItem {
+    private Long value;
+    private double changeRate;
+}
