+++ src/main/java/com/modive/adminservice/external/user/dto/res/UCTotalDeviceResData.java
@@ -1,0 +1,14 @@
+package com.modive.adminservice.external.user.dto.res;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UCTotalDeviceResData {
+    private UCTotalCntAndRateItem totalCarCount;
+}
